Move AutoDetectionDraft regression test into the Issue186Test structure
Relocates the $schema-override regression coverage out of ArrayContainsTest
into tests/Issues/Issue/Issue186Test.php with its fixture under
tests/Schema/Issues/186/, matching this repo's convention for issue-driven
regression tests (see Issue179Test, Issue169Test, etc.) instead of living
alongside the general contains/minContains/maxContains behavior tests.
