+++ tests/Schema/Issues/186/AutoDetectionEmbeddedResourceSchemaOverride.json
